[Threading] Dotnet pool second pass
