--- sources/core/Stride.Core/Threading/LowLevelSpinWaiter.cs
@@ -1,12 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Diagnostics;
-
-namespace System.Threading
-{
-    using Reflection;
-
-
-}
