+++ sources/core/Stride.Core/Threading/ThreadPool.HillClimbing.cs
@@ -5,7 +5,7 @@
 using System.Diagnostics;
 
 using System;
-namespace SubSystem.Threading
+namespace Stride.Core.Threading
 {
     public partial class ThreadPool
     {
